[Chore #162461660] Add distance matrix in home page
