+++ ui/src/index.html
@@ -28,8 +28,5 @@
   </footer>
   <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDu_9x-afqT01pFFB3vc4cDRObiSVIen1I&callback=initMap"
   async defer></script>
-  <!--<script src="./js/main.js"></script>
-  <script src="./js/quote.js"></script>-->
-  <script src="./utils/defaultInitMap.js"></script> 
 </body>
 </html>
